+++ src/oberwolfach/peekInductor.jl
@@ -1,0 +1,11 @@
+# This file is part of MORWiki. License is MIT: https://spdx.org/licenses/MIT.html
+
+"""
+    PeekInductor() <: Benchmark
+
+!!! compat "MORWiki v0.3.4"
+    This data set requires MORWiki version 0.3.4 or later.
+"""
+struct PeekInductor <: Benchmark end
+
+@associate PeekInductor() "peekInductor_n1434m1q1"